+++ openbb_terminal/stocks/fundamental_analysis/fa_api.py
@@ -9,7 +9,7 @@ from openbb_terminal.helper_classes import ModelsNamespace as _models
 from .eclect_us_view import display_analysis as analysis
 from .business_insider_view import display_management as mgmt
 from .finviz_view import display_screen_data as data
-from .financial_modeling_prep.fmp_view import valinvest_score as score
+from .fmp_view import valinvest_score as score
 from .yahoo_finance_view import display_info as info
 from .yahoo_finance_view import display_shareholders as shrs
 from .yahoo_finance_view import display_sustainability as sust
@@ -26,9 +26,6 @@ from .av_view import display_earnings as earnings
 from .av_view import display_fraud as fraud
 from .dcf_view import CreateExcelFA as dcf
 from .market_watch_view import display_sean_seah_warnings as warnings
-
-# Submenus
-from .financial_modeling_prep import fmp_api as fmp
 
 # Models
 models = _models(os.path.abspath(os.path.dirname(__file__)))
