+++ openbb_terminal/stocks/fundamental_analysis/fa_controller.py
@@ -30,9 +30,6 @@ from openbb_terminal.stocks.fundamental_analysis import (
     market_watch_view,
     yahoo_finance_view,
     polygon_view,
-)
-from openbb_terminal.stocks.fundamental_analysis.financial_modeling_prep import (
-    fmp_controller,
     fmp_view,
 )
 
@@ -49,6 +46,12 @@ class FundamentalAnalysisController(StockBaseController):
         "load",
         "analysis",
         "score",
+        "profile",
+        "quote",
+        "enterprise",
+        "metrics",
+        "ratios",
+        "growth",
         "warnings",
         "data",
         "income",
@@ -71,13 +74,11 @@ class FundamentalAnalysisController(StockBaseController):
         "earnings",
         "fraud",
         "dcf",
+        "dcfc",
         "mktcap",
         "dupont",
     ]
 
-    CHOICES_MENUS = [
-        "fmp",
-    ]
     PATH = "/stocks/fa/"
 
     def __init__(
@@ -113,8 +114,15 @@ class FundamentalAnalysisController(StockBaseController):
         mt.add_cmd("mgmt", "Business Insider")
         mt.add_cmd("analysis", "Elect")
         mt.add_cmd("score", "FMP")
+        mt.add_cmd("profile", "FMP")
+        mt.add_cmd("quote", "FMP")
+        mt.add_cmd("enterprise", "FMP")
+        mt.add_cmd("metrics", "FMP")
+        mt.add_cmd("ratios", "FMP")
+        mt.add_cmd("growth", "FMP")
         mt.add_cmd("warnings", "Market Watch")
         mt.add_cmd("dcf", "Stockanalysis")
+        mt.add_cmd("dcfc", "FMP")
         mt.add_cmd("info", "Yahoo Finance")
         mt.add_cmd("mktcap", "Yahoo Finance")
         mt.add_cmd("shrs", "Yahoo Finance", not self.suffix)
@@ -124,17 +132,16 @@ class FundamentalAnalysisController(StockBaseController):
         mt.add_cmd("splits", "Yahoo Finance", not self.suffix)
         mt.add_cmd("web", "Yahoo Finance", not self.suffix)
         mt.add_cmd("hq", "Yahoo Finance", not self.suffix)
-        mt.add_cmd("income", "Alpha Vantage / Polygon")
-        mt.add_cmd("balance", "Alpha Vantage / Polygon")
+        mt.add_cmd("income", "Alpha Vantage / Polygon / FMP")
+        mt.add_cmd("balance", "Alpha Vantage / Polygon / FMP")
         mt.add_cmd("overview", "Alpha Vantage")
         mt.add_cmd("key", "Alpha Vantage")
-        mt.add_cmd("cash", "Alpha Vantage")
+        mt.add_cmd("cash", "Alpha Vantage / FMP")
         mt.add_cmd("earnings", "Alpha Vantage")
         mt.add_cmd("fraud", "Alpha Vantage")
         mt.add_cmd("dupont", "Alpha Vantage")
         mt.add_raw("\n")
         mt.add_info("_sources_")
-        mt.add_menu("fmp")
         console.print(text=mt.menu_text, menu="Stocks - Fundamental Analysis")
 
     def custom_reset(self):
@@ -226,6 +233,251 @@ class FundamentalAnalysisController(StockBaseController):
         )
         if ns_parser:
             fmp_view.valinvest_score(self.ticker)
+
+    @log_start_end(log=logger)
+    def call_profile(self, other_args: List[str]):
+        """Process profile command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="profile",
+            description="""
+                    Prints information about, among other things, the industry, sector exchange and company
+                    description. The following fields are expected: Address, Beta, Ceo, Changes, Cik, City
+                    Company name, Country, Currency, Cusip, Dcf, Dcf diff, Default image, Description,
+                    Exchange, Exchange short name, Full time employees, Image, Industry, Ipo date, Isin,
+                    Last div, Mkt cap, Phone, Price, Range, Sector, State, Symbol, Vol avg, Website, Zip.
+                    [Source: Financial Modeling Prep]
+                """,
+        )
+        ns_parser = parse_known_args_and_warn(parser, other_args)
+        if ns_parser:
+            fmp_view.display_profile(self.ticker)
+
+    @log_start_end(log=logger)
+    def call_quote(self, other_args: List[str]):
+        """Process quote command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="quote",
+            description="""
+                    Prints actual information about the company which is, among other things, the day high,
+                    market cap, open and close price and price-to-equity ratio. The following fields are
+                    expected: Avg volume, Change, Changes percentage, Day high, Day low, Earnings
+                    announcement, Eps, Exchange, Market cap, Name, Open, Pe, Previous close, Price, Price
+                    avg200, Price avg50, Shares outstanding, Symbol, Timestamp, Volume, Year high, and Year
+                    low. [Source: Financial Modeling Prep]
+                """,
+        )
+        ns_parser = parse_known_args_and_warn(parser, other_args)
+        if ns_parser:
+            fmp_view.display_quote(self.ticker)
+
+    @log_start_end(log=logger)
+    def call_enterprise(self, other_args: List[str]):
+        """Process income command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="enterprise",
+            description="""
+                    Prints stock price, number of shares, market capitalization and
+                    enterprise value over time. The following fields are expected: Add total debt,
+                    Enterprise value, Market capitalization, Minus cash and cash equivalents, Number
+                    of shares, Stock price, and Symbol. [Source: Financial Modeling Prep]
+                """,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            action="store",
+            dest="limit",
+            type=check_positive,
+            default=1,
+            help="Limit of latest years/quarters.",
+        )
+        parser.add_argument(
+            "-q",
+            "--quarter",
+            action="store_true",
+            default=False,
+            dest="b_quarter",
+            help="Quarter fundamental data flag.",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if ns_parser:
+            fmp_view.display_enterprise(
+                ticker=self.ticker,
+                number=ns_parser.limit,
+                quarterly=ns_parser.b_quarter,
+                export=ns_parser.export,
+            )
+
+    @log_start_end(log=logger)
+    def call_metrics(self, other_args: List[str]):
+        """Process metrics command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="metrics",
+            description="""
+                    Prints a list of the key metrics of a company over time. This can be either
+                    quarterly or annually. This includes, among other things, Return on Equity (ROE),
+                    Working Capital, Current Ratio and Debt to Assets. The following fields are expected:
+                    Average inventory, Average payables, Average receivables, Book value per share, Capex
+                    per share, Capex to depreciation, Capex to operating cash flow, Capex to revenue, Cash
+                    per share, Current ratio, Days of inventory on hand, Days payables outstanding, Days
+                    sales outstanding, Debt to assets, Debt to equity, Dividend yield, Earnings yield,
+                    Enterprise value, Enterprise value over EBITDA, Ev to free cash flow, Ev to operating
+                    cash flow, Ev to sales, Free cash flow per share, Free cash flow yield, Graham net net,
+                    Graham number, Income quality, Intangibles to total assets, Interest debt per share,
+                    Inventory turnover, Market cap, Net current asset value, Net debt to EBITDA, Net income
+                    per share, Operating cash flow per share, Payables turnover, Payout ratio, Pb ratio, Pe
+                    ratio, Pfcf ratio, Pocf ratio, Price to sales ratio, Ptb ratio, Receivables turnover,
+                    Research and development to revenue, Return on tangible assets, Revenue per share,
+                    Roe, Roic, Sales general and administrative to revenue, Shareholders equity per
+                    share, Stock based compensation to revenue, Tangible book value per share, and Working
+                    capital. [Source: Financial Modeling Prep]
+                """,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            action="store",
+            dest="limit",
+            type=check_positive,
+            default=1,
+            help="Limit of latest years/quarters.",
+        )
+        parser.add_argument(
+            "-q",
+            "--quarter",
+            action="store_true",
+            default=False,
+            dest="b_quarter",
+            help="Quarter fundamental data flag.",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if ns_parser:
+            fmp_view.display_key_metrics(
+                ticker=self.ticker,
+                number=ns_parser.limit,
+                quarterly=ns_parser.b_quarter,
+                export=ns_parser.export,
+            )
+
+    @log_start_end(log=logger)
+    def call_ratios(self, other_args: List[str]):
+        """Process ratios command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="ratios",
+            description="""
+                    Prints in-depth ratios of a company over time. This can be either quarterly or
+                    annually. This contains, among other things, Price-to-Book Ratio, Payout Ratio and
+                    Operating Cycle. The following fields are expected: Asset turnover, Capital expenditure
+                    coverage ratio, Cash conversion cycle, Cash flow coverage ratios, Cash flow to debt
+                    ratio, Cash per share, Cash ratio, Company equity multiplier, Current ratio, Days of
+                    inventory outstanding, Days of payables outstanding, Days of sales outstanding, Debt
+                    equity ratio, Debt ratio, Dividend paid and capex coverage ratio, Dividend payout ratio,
+                    Dividend yield, Ebit per revenue, Ebt per ebit, Effective tax rate, Enterprise value
+                    multiple, Fixed asset turnover, Free cash flow operating cash flow ratio, Free cash
+                    flow per share, Gross profit margin, Inventory turnover, Long term debt to
+                    capitalization, Net income per EBT, Net profit margin, Operating cash flow per share,
+                    Operating cash flow sales ratio, Operating cycle, Operating profit margin, Payables
+                    turnover, Payout ratio, Pretax profit margin, Price book value ratio, Price cash flow
+                    ratio, Price earnings ratio, Price earnings to growth ratio, Price fair value,
+                    Price sales ratio, Price to book ratio, Price to free cash flows ratio, Price to
+                    operating cash flows ratio, Price to sales ratio, Quick ratio, Receivables turnover,
+                    Return on assets, Return on capital employed, Return on equity, Short term coverage
+                    ratios, and Total debt to capitalization. [Source: Financial Modeling Prep]
+                """,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            action="store",
+            dest="limit",
+            type=check_positive,
+            default=1,
+            help="Limit of latest years/quarters.",
+        )
+        parser.add_argument(
+            "-q",
+            "--quarter",
+            action="store_true",
+            default=False,
+            dest="b_quarter",
+            help="Quarter fundamental data flag.",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if ns_parser:
+            fmp_view.display_financial_ratios(
+                ticker=self.ticker,
+                number=ns_parser.limit,
+                quarterly=ns_parser.b_quarter,
+                export=ns_parser.export,
+            )
+
+    @log_start_end(log=logger)
+    def call_growth(self, other_args: List[str]):
+        """Process growth command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="growth",
+            description=""" Prints the growth of several financial statement items and ratios over
+                time. This can be either annually and quarterly. These are, among other things, Revenue
+                Growth (3, 5 and 10 years), inventory growth and operating cash flow growth (3, 5 and 10
+                years). The following fields are expected: Asset growth, Book valueper share growth, Debt
+                growth, Dividendsper share growth, Ebit growth, Eps diluted growth, Eps growth, Five y
+                dividendper share growth per share, Five y net income growth per share, Five y operating c
+                f growth per share, Five y revenue growth per share, Five y shareholders equity growth per
+                share, Free cash flow growth, Gross profit growth, Inventory growth, Net income growth,
+                Operating cash flow growth, Operating income growth, Rd expense growth, Receivables growth,
+                Revenue growth, Sga expenses growth, Ten y dividendper share growth per share, Ten y net
+                income growth per share, Ten y operating c f growth per share, Ten y revenue growth per
+                share, Ten y shareholders equity growth per share, Three y dividendper share growth per
+                share, Three y net income growth per share, Three y operating c f growth per share, Three y
+                revenue growth per share, Three y shareholders equity growth per share, Weighted average
+                shares diluted growth, and Weighted average shares growth [Source: Financial Modeling Prep]
+                """,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            action="store",
+            dest="limit",
+            type=check_positive,
+            default=1,
+            help="Limit of latest years/quarters.",
+        )
+        parser.add_argument(
+            "-q",
+            "--quarter",
+            action="store_true",
+            default=False,
+            dest="b_quarter",
+            help="Quarter fundamental data flag.",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if ns_parser:
+            fmp_view.display_financial_statement_growth(
+                ticker=self.ticker,
+                number=ns_parser.limit,
+                quarterly=ns_parser.b_quarter,
+                export=ns_parser.export,
+            )
 
     @log_start_end(log=logger)
     def call_info(self, other_args: List[str]):
@@ -484,7 +736,7 @@ class FundamentalAnalysisController(StockBaseController):
 
     @log_start_end(log=logger)
     def call_key(self, other_args: List[str]):
-        """Process overview command."""
+        """Process key command."""
         parser = argparse.ArgumentParser(
             add_help=False,
             formatter_class=argparse.ArgumentDefaultsHelpFormatter,
@@ -539,11 +791,35 @@ class FundamentalAnalysisController(StockBaseController):
             help="Quarter fundamental data flag.",
         )
         parser.add_argument(
+            "-r",
+            "--ratios",
+            action="store_true",
+            default=False,
+            dest="ratios",
+            help="Shows percentage change of values.",
+        )
+        parser.add_argument(
+            "-p",
+            "--plot",
+            action="store",
+            nargs="+",
+            default=None,
+            choices=["Revenue", "Cost_of_revenue", "Gross_profit", "Gross_profit_ratio", "Research_and_development_expenses",
+                     "General_and_administrative_expenses", "Selling and_marketing_expenses", "Selling_and_marketing_expenses",
+                     "Selling_general_and_administrative_expenses", "Other_expenses", "Operating_expenses", "Cost_and_expenses",
+                     "Interest_income", "Interest_expense", "Depreciation_and_amortization", "Ebitda", "Ebitdaratio", "Operating_income",
+                     "Operating_income_ratio", "Total_other_income_expenses_net", "Income_before_tax", "Income_before_tax_ratio",
+                     "Income_tax_expense", "Net_income", "Net_income_ratio", "Eps", "Epsdiluted", "Weighted_average_shs_out",
+                     "Weighted_average_shs_out_dil"],
+            dest="plot",
+            help="Rows to plot."
+        )
+        parser.add_argument(
             "-s",
             "--source",
             help="Source to get fundamentals from",
             default="av",
-            choices=["polygon", "av"],
+            choices=["polygon", "av", "fmp"],
             dest="source",
         )
         ns_parser = parse_known_args_and_warn(
@@ -563,6 +839,15 @@ class FundamentalAnalysisController(StockBaseController):
                     financial="income",
                     limit=ns_parser.limit,
                     quarterly=ns_parser.b_quarter,
+                    export=ns_parser.export,
+                )
+            elif ns_parser.source == "fmp":
+                fmp_view.display_income_statement(
+                    ticker=self.ticker,
+                    number=ns_parser.limit,
+                    quarterly=ns_parser.b_quarter,
+                    ratios=ns_parser.ratios,
+                    plot=ns_parser.plot,
                     export=ns_parser.export,
                 )
 
@@ -612,7 +897,7 @@ class FundamentalAnalysisController(StockBaseController):
             "--source",
             help="Source to get fundamentals from",
             default="av",
-            choices=["polygon", "av"],
+            choices=["polygon", "av", "fmp"],
             dest="source",
         )
         ns_parser = parse_known_args_and_warn(
@@ -631,6 +916,13 @@ class FundamentalAnalysisController(StockBaseController):
                     ticker=self.ticker,
                     financial="balance",
                     limit=ns_parser.limit,
+                    quarterly=ns_parser.b_quarter,
+                    export=ns_parser.export,
+                )
+            elif ns_parser.source == "fmp":
+                fmp_view.display_balance_sheet(
+                    ticker=self.ticker,
+                    number=ns_parser.limit,
                     quarterly=ns_parser.b_quarter,
                     export=ns_parser.export,
                 )
@@ -674,16 +966,32 @@ class FundamentalAnalysisController(StockBaseController):
             dest="b_quarter",
             help="Quarter fundamental data flag.",
         )
+        parser.add_argument(
+            "-s",
+            "--source",
+            help="Source to get fundamentals from",
+            default="av",
+            choices=["av", "fmp"],
+            dest="source",
+        )
         ns_parser = parse_known_args_and_warn(
             parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
         )
         if ns_parser:
-            av_view.display_cash_flow(
-                ticker=self.ticker,
-                limit=ns_parser.limit,
-                quarterly=ns_parser.b_quarter,
-                export=ns_parser.export,
-            )
+            if ns_parser.source == "av":
+                av_view.display_cash_flow(
+                    ticker=self.ticker,
+                    limit=ns_parser.limit,
+                    quarterly=ns_parser.b_quarter,
+                    export=ns_parser.export,
+                )
+            elif ns_parser.source == "fmp":
+                fmp_view.display_cash_flow(
+                    ticker=self.ticker,
+                    number=ns_parser.limit,
+                    quarterly=ns_parser.b_quarter,
+                    export=ns_parser.export,
+                )
 
     @log_start_end(log=logger)
     def call_earnings(self, other_args: List[str]):
@@ -900,6 +1208,47 @@ class FundamentalAnalysisController(StockBaseController):
             dcf.create_workbook()
 
     @log_start_end(log=logger)
+    def call_dcfc(self, other_args: List[str]):
+        """Process dcfc command"""
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+            prog="dcfc",
+            description="""
+                    Prints the discounted cash flow of a company over time including the DCF of today. The
+                    following fields are expected: DCF, Stock price, and Date. [Source: Financial Modeling
+                    Prep]
+                """,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            action="store",
+            dest="limit",
+            type=check_positive,
+            default=1,
+            help="Limit of latest years/quarters.",
+        )
+        parser.add_argument(
+            "-q",
+            "--quarter",
+            action="store_true",
+            default=False,
+            dest="b_quarter",
+            help="Quarter fundamental data flag.",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if ns_parser:
+            fmp_view.display_discounted_cash_flow(
+                ticker=self.ticker,
+                number=ns_parser.limit,
+                quarterly=ns_parser.b_quarter,
+                export=ns_parser.export,
+            )
+
+    @log_start_end(log=logger)
     def call_warnings(self, other_args: List[str]):
         """Process warnings command."""
         parser = argparse.ArgumentParser(
@@ -928,17 +1277,6 @@ class FundamentalAnalysisController(StockBaseController):
             market_watch_view.display_sean_seah_warnings(
                 ticker=self.ticker, debug=ns_parser.b_debug
             )
-
-    @log_start_end(log=logger)
-    def call_fmp(self, _):
-        """Process fmp command."""
-        self.queue = self.load_class(
-            fmp_controller.FinancialModelingPrepController,
-            self.ticker,
-            self.start,
-            self.interval,
-            self.queue,
-        )
 
 
 def key_metrics_explained(other_args: List[str]):
