+++ openbb_terminal/stocks/fundamental_analysis/fmp_model.py
@@ -1,7 +1,7 @@
 """ Financial Modeling Prep Model"""
 __docformat__ = "numpy"
 import logging
-from typing import Optional
+from typing import Optional, Union
 
 from datetime import datetime
 from requests.exceptions import HTTPError
@@ -50,7 +50,18 @@ def get_score(ticker: str) -> Optional[np.number]:
 
 @log_start_end(log=logger)
 def get_profile(ticker: str) -> pd.DataFrame:
-    """Get ticker profile from FMP"""
+    """Get ticker profile from FMP
+
+    Parameters
+    ----------
+    ticker : str
+        Stock ticker
+
+    Returns
+    ----------
+    pd.DataFrame:
+        Dataframe of ticker profile
+    """
     df = pd.DataFrame()
 
     try:
@@ -66,7 +77,18 @@ def get_profile(ticker: str) -> pd.DataFrame:
 
 @log_start_end(log=logger)
 def get_quote(ticker) -> pd.DataFrame:
-    """Gets ticker quote from FMP"""
+    """Gets ticker quote from FMP
+
+    Parameters
+    ----------
+    ticker : str
+        Stock ticker
+
+    Returns
+    ----------
+    pd.DataFrame:
+        Dataframe of ticker quote
+    """
 
     df_fa = pd.DataFrame()
 
@@ -178,7 +200,7 @@ def get_dcf(ticker: str, number: int, quarterly: bool = False) -> pd.DataFrame:
 
 
 @log_start_end(log=logger)
-def get_income(ticker: str, number: int, quarterly: bool = False) -> pd.DataFrame:
+def get_income(ticker: str, number: int, quarterly: bool = False, ratios: bool = False) -> list[Union[pd.DataFrame, pd.DataFrame]]:
     """Get income statements
 
     Parameters
@@ -189,14 +211,17 @@ def get_income(ticker: str, number: int, quarterly: bool = False) -> pd.DataFram
         Number to get
     quarterly : bool, optional
         Flag to get quarterly data, by default False
+    ratios: bool
+        Shows percentage change, by default False
 
     Returns
     -------
-    pd.DataFrame
-        Dataframe of income statements
+    list
+        List of two Dataframes first, is income statement table formatted and the second is for plotting
     """
 
     df_fa = pd.DataFrame()
+    df_fa_c = pd.DataFrame()
 
     try:
         if quarterly:
@@ -205,7 +230,7 @@ def get_income(ticker: str, number: int, quarterly: bool = False) -> pd.DataFram
             )
         else:
             df_fa = fa.income_statement(ticker, cfg.API_KEY_FINANCIALMODELINGPREP)
-        df_fa = clean_metrics_df(df_fa, num=number)
+        df_fa_c = clean_metrics_df(df_fa, num=number)
     # Invalid API Keys
     except ValueError as e:
         console.print(e)
@@ -213,7 +238,23 @@ def get_income(ticker: str, number: int, quarterly: bool = False) -> pd.DataFram
     except HTTPError as e:
         console.print(e)
 
-    return df_fa
+    if ratios:
+        types = df_fa.copy().applymap(lambda x: type(x) == int or type(x) == float).all(axis=1)
+        valid = []
+        i = 0
+        for row in types:
+            if row:
+                valid.append(i)
+            i += 1
+        df_fa_pc = df_fa.iloc[valid].pct_change(axis="columns", periods=-1).fillna(0)
+        j = 0
+        for i in valid:
+            df_fa.iloc[i] = df_fa_pc.iloc[j]
+            j += 1
+        df_fa = df_fa.iloc[:, 0:number]
+        df_fa_c = clean_metrics_df(df_fa, num=number)
+
+    return [df_fa_c, df_fa]
 
 
 @log_start_end(log=logger)
